routing: add null check for config dir deletion
The missing null check was causing static routing examples
to exit with error during clean up as config files are not
created for static routing.

+++ nest/routing/routing_helper.py
@@ -331,7 +331,7 @@ class RoutingHelper:
                         pass
 
         # Delete config directory
-        if path.isdir(self.conf_dir):
+        if self.conf_dir is not None and path.isdir(self.conf_dir):
             rmtree(self.conf_dir)
 
         # Change ownership of log files to current user
